requirements.txt
streamlit
pandas
openpyxl
plotly
fpdf2
python-docx